pyproject.toml
[project]
name = "muxpack"
version = "0.1.0"
description = "Tools to handle multiplex network data more easily"
readme = "README.md"
authors = [
    { name = "Edwin de Jonge", email = "edwindjonge@gmail.com" },
    { name = "Jan van der Laan", email = "djvanderlaan@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.4",
    "ibis-framework[duckdb]>=12.0.0",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
    "pyarrow-hotfix>=0.7",
    "scipy>=1.17.1",
]

[project.scripts]
muxpack = "muxpack:main"

[build-system]
requires = ["uv_build>=0.10.6,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.4",
]
docs = [
    "sphinx>=9.1.0",
]

[tool.ruff.lint]
extend-select = ["B"]
